Revamp play bottom sheet
