+++ lib/src/view/play/quick_game_matrix.dart
@@ -10,7 +10,7 @@ import 'package:lichess_mobile/src/network/connectivity.dart';
 import 'package:lichess_mobile/src/styles/styles.dart';
 import 'package:lichess_mobile/src/utils/l10n_context.dart';
 import 'package:lichess_mobile/src/view/game/game_screen.dart';
-import 'package:lichess_mobile/src/view/play/create_custom_game_screen.dart';
+import 'package:lichess_mobile/src/view/play/play_bottom_sheet.dart';
 import 'package:lichess_mobile/src/view/play/playban.dart';
 
 const _kMatrixSpacing = 8.0;
@@ -122,7 +122,12 @@ class _SectionChoices extends ConsumerWidget {
                 onTap:
                     isOnline
                         ? () {
-                          Navigator.of(context).push(CreateCustomGameScreen.buildRoute(context));
+                          showModalBottomSheet<void>(
+                            context: context,
+                            builder: (context) {
+                              return const PlayBottomSheet();
+                            },
+                          );
                         }
                         : null,
               ),
